Move getPopupUrl, setUrlContext to src/shared
